SystemCode/stock/requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
dialogflow==1.1.1
docopt==0.6.2
Flask==2.0.2
google-api-core==1.31.3
google-auth==1.35.0
google-cloud==0.34.0
google-cloud-aiplatform==1.6.0
google-cloud-bigquery==2.28.1
google-cloud-core==2.1.0
google-cloud-dialogflow==2.9.1
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
packaging==21.0
pandas==1.1.5
Pillow==8.4.0
pipreqs==0.4.11
proto-plus==1.19.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
yarg==0.1.9
zipp==3.6.0
